OGM-860 Move DocumentHelpers to more sensible place
